Add basic error reporter
It's a role, and for now just does one simple fatal message, styled to
look like the typical Perl 6 compilation errors (though the ===SORRY!===
was purposely colored differently to distinguish from actual compiler
errors).

Typed errors should be the next goal, so that messages aren't only
arbitrary strings. The eventual goal with the error reporter, by the
way, is to develop similar facilities for P6 grammars that the NQP
grammars in Rakudo have. Who knows when the error handler is split off
from this? :)