Remove scope locking from PodScope
This is, I think, a relic from an older model of handling pod scopes,
where the =config directive didn't get its own scope from <block>. I
think. Point is, there's no real issue of =config lines messing stuff up
in other scopes. The worst thing is that, when implemented, stuff
applied in a =config has to be transfered from [*-1] to [*-2], which
isn't that big of a deal.