Fix up configuration accumulation
This makes it so that, at a certain level, @*POD_BLOCKS is used
entirely, skipping 'make' and thus leaving a few things empty, such as
the directive methods. I'm still not sure if this is the best choice,
but it seems the most natural fit so far. I'll investigate trying to use
'make' at some point, and see what that would look like.

The main reason for using @*POD_BLOCKS as a kind of @!BLOCKS (in
Perl6::World) is because certain configuration options affect parsing,
and if we didn't grab them for use when encountered, we'd have to
re-parse things. The alternative PodScope method, which separately
tracked such info, was tossed because it was turning out to be a second
version of the existing Pod6 objects.