Create my first test file, block_types.t
This tests all the standard, not-semantic block types, except for
=table. Some tests have been manually skipped for now (since there's no
automatic fudger available for general P6 use, perhaps I should write
one...), in the interest of making the test file run completely through.

This file is written specifically for SUPERNOVA, since I figured trying
to write things in the way that they would be in roast/ was going to be
too much of a pain to be worth it right now.

Also, a couple of minor fixes to lib/ were made while writing this.