Concatenate adjacent plaintext blocks
This reduces the number of Pod6::Text::Plain blocks that ultimately get
placed. Note that formatting code blocks, when (re)implemented, won't
get concatenated.

While I am misusing the name 'append' for the Str candidate a tiny bit,
I'll point out that I always meant to make a method like this. Should
probably be replaced with using 'push' anyway, but for now I figured it
might be easier to not worry with append.