Switch to using Grammar::Parsefail for errors
This is both a bit of dogfooding (trying this switch the first time led
to the subrule test, after all), and a bit of not replicating code. Of
course in the move to core this will be using rakudo's existing error
stuff, this just lets us develop the error side in advance.