Add formatting code support
This is a big commit, and I'm not 100% happy with out it's turned out,
but it does work. The @*FORMAT_CODES exists just to make use of
preserve_spaces and such (not yet, but most likely next commit).

The rationale for this, by the way, is to be an easy way to signal
various ways of parsing stuff. Especially useful for custom blocks. In
the future I may want to switch to using =config options (with a set of
defaults), and generate blocks in a more traditional 'make/.ast'
fashion. Not sure how custom blocks may be able to work out their
default behaviors then.