Parse and die on reserved formatting codes.
The parser will now die if it looks like you started a formatting code
that's been reserved. The reason for doing this, instead of just not
recognizing them as codes (previous behavior), is so that any future
defined codes won't cause problems for people when their
wasn't-a-formatting-code text suddenly is interpreted differently.