Fix P<>.text not meant to work
Since .text is meant to be 'display text' for each of the classes, and
P<> is designed to pull text from an external source, P<>.text is
currently set to die. As I envision it, renderers are currently
responsible for retrieving the text, and at the point renderers are
working, .text shouldn't need to be set to the proper value. (That could
be arranged if needed, though.)