Rework how the AST thing works
I really wasn't happy with the "as you go" generation of blocks and
such, and thus rewrote it to hold the important parse-time info in a
hash and an array of margins. In the process a new file was added for a
separate, fake version of Perl6::World, in the interest of easing
transition to core.

Not all tests pass, but all the files do run.