Make tests pass with rewrite of AST stuff
This included a quick fix to Defn to un-skip its tests. The way it
turned out suggests specialized grammar rules for defn might be
warranted. After all, its parsing is meant to be slightly different from
the normal procedure. Probably do that later, but for now this works.

Also changed configuration options to be supplied to the objects via a
:config option, instead of taking any random nameds. This will work
better to allow configuration options that happen to match named params
to new(), and feels nicer to do in general. (Classes don't get config
options since the rewrite, but it should be put back in soon.)