Add test file for tables.
Currently passes, only because table blocks do get parsed, but as a
normal block.

The tests of course miss out on a wide variety of possible tables, but I
think the tests here provide a good mixture of possibilities for a
start. You'll also notice the long comment at top describing the rules
for tables as I currently see them, since S26 is lacking on describing
them.