Tank Pod6 Config class
This was part of an old way of handling AST stuff, and was made
irrelevant by the refactor. So =config settings are handled, but not
inserted into the AST as an object, so only used configuration options
will end up showing up, attached to what uses them.

We could add this object back for sure, esp. if/when certain modules
want or need otherwise-unused configuration specifications. The removal
was to keep a possibly confusing piece of cruft from hanging around, not
so much to say it'll never be needed again.